PUSHING SOURCE CODE TO GIT

+++ Jenkinsfile
@@ -15,9 +15,19 @@ pipeline {
                 script {
                     sh '''
                     ImageId=$(docker images -q)
-                    if [ -n "$ImageId" ]; then
+                    ContainerId=$(docker ps -q)
+                    if [ -n "$ImageId" ] && [ -n "$ContainerId" ] ; then
+                        echo "STOPPING THE CONTAINER: $ContainerId"
+                        docker stop $ContainerId
                         echo "IMAGE ID IS: $ImageId"
                         docker rmi -f $ImageId
+
+                        if [ $? -eq 0 ]; then
+                            echo "REMOVING THE CONATINER: $ContainerId"
+                            docker rm $ContainerId  
+                        fi
+                    else
+                        echo "NO RUNNING CONTAINER & IMAGES!!!"
                     fi
                         
                     docker build -t abhishekswarnakar/front-end_image:${BUILD_NUMBER} -f ./Front-End/Dockerfile ./Front-End
@@ -30,23 +40,6 @@ pipeline {
         stage('Run front-end Docker image') {
             steps {
                 script {
-
-                    sh '''
-                    ContainerId=$(docker ps -q)
-                    if [ -n "$ContainerId" ]; then
-                        echo "STOPPING THE CONTAINER: $ContainerId"
-                        docker stop $ContainerId
-                        
-
-                        if [ $? -eq 0 ]; then
-                            echo "REMOVING THE CONATINER: $ContainerId"
-                            docker rm $ContainerId  
-                        fi
-                    else
-                        echo "NO RUNNING CONTAINER & IMAGES!!!"
-                    fi
-                    '''
-
                     // Run new container with a unique name
                     sh 'docker run -d --name front-end_container_${BUILD_NUMBER} -p 8000:8000 abhishekswarnakar/front-end_image:${BUILD_NUMBER}'
                 }
